Rename GMP 21.10 to GMP 22.04
